clean up note_deprecation
